Add support for expressions, booleans, values, stop (return), if
This commit extends the language to be able to run the tree program.
The tree program use mathematical expression, and if test, boolean
expressions, and stop (to return from a procedure).

Right now the tree program is written out directly as an AST.
The next step is to extend the parser to support these new language
features.